requirements.txt
# Angel Cloud / LogiBot Python Dependencies

# Firebase Admin SDK for Firestore
firebase-admin>=6.2.0

# Google Sheets API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Weaviate Vector Database
weaviate-client>=3.0.0
